fix(helperMiniAOD): fix empty charge branches in treeR (wrong datatype)

+++ EventAnalyzer/test/helperMiniAOD.C
@@ -906,32 +906,32 @@ void setupVariables2(BaseHandler* handler,bool isMC = false)
   handler->addEventVariable("TOTALISOBASICTAUS", new EventVariableObjectVariableVector<double>("TOTALISO","basicTaus"));
 
   handler->addEventVariable("NGOODELECTRONS", new EventVariableN("NGOODELECTRONS","goodElectrons"));
-  handler->addEventVariable("QGOODELECTRONS", new EventVariableObjectVariableVector<double>("CHARGE","goodElectrons"));
+  handler->addEventVariable("QGOODELECTRONS", new EventVariableObjectVariableVector<int>("CHARGE","goodElectrons"));
   handler->addEventVariable("PTGOODELECTRONS", new EventVariableObjectVariableVector<double>("PT","goodElectrons"));
   handler->addEventVariable("ETAGOODELECTRONS", new EventVariableObjectVariableVector<double>("ETA","goodElectrons"));
   handler->addEventVariable("fakeRoleGOODELECTRONS", new EventVariableObjectVariableVector<int>("fakeRole","goodElectrons"));
   
   handler->addEventVariable("NGOODMUONS",new EventVariableN("NGOODMUONS","goodMuons"));
-  handler->addEventVariable("QGOODMUONS", new EventVariableObjectVariableVector<double>("CHARGE","goodMuons"));
+  handler->addEventVariable("QGOODMUONS", new EventVariableObjectVariableVector<int>("CHARGE","goodMuons"));
   handler->addEventVariable("PTGOODMUONS", new EventVariableObjectVariableVector<double>("PT","goodMuons"));
   handler->addEventVariable("ETAGOODMUONS", new EventVariableObjectVariableVector<double>("ETA","goodMuons"));
   handler->addEventVariable("fakeRoleGOODMUONS", new EventVariableObjectVariableVector<int>("fakeRole","goodMuons"));
 
   handler->addEventVariable("NGOODELECTRONSLOWPT", new EventVariableN("NGOODELECTRONSLOWPT","goodElectronsLowPt"));
-  handler->addEventVariable("QGOODELECTRONSLOWPT", new EventVariableObjectVariableVector<double>("CHARGE","goodElectronsLowPt"));
+  handler->addEventVariable("QGOODELECTRONSLOWPT", new EventVariableObjectVariableVector<int>("CHARGE","goodElectronsLowPt"));
   handler->addEventVariable("PTGOODELECTRONSLOWPT", new EventVariableObjectVariableVector<double>("PT","goodElectronsLowPt"));
   handler->addEventVariable("ETAGOODELECTRONSLOWPT", new EventVariableObjectVariableVector<double>("ETA","goodElectronsLowPt"));
   handler->addEventVariable("fakeRoleGOODELECTRONSLOWPT", new EventVariableObjectVariableVector<int>("fakeRole","goodElectronsLowPt"));
   
   handler->addEventVariable("NGOODMUONSLOWPT",new EventVariableN("NGOODMUONSLOWPT","goodMuonsLowPt"));
-  handler->addEventVariable("QGOODMUONSLOWPT", new EventVariableObjectVariableVector<double>("CHARGE","goodMuonsLowPt"));
+  handler->addEventVariable("QGOODMUONSLOWPT", new EventVariableObjectVariableVector<int>("CHARGE","goodMuonsLowPt"));
   handler->addEventVariable("PTGOODMUONSLOWPT", new EventVariableObjectVariableVector<double>("PT","goodMuonsLowPt"));
   handler->addEventVariable("ETAGOODMUONSLOWPT", new EventVariableObjectVariableVector<double>("ETA","goodMuonsLowPt"));
   handler->addEventVariable("fakeRoleGOODMUONSLOWPT", new EventVariableObjectVariableVector<int>("fakeRole","goodMuonsLowPt"));
 
   
   handler->addEventVariable("NGOODTAUS", new EventVariableN("NGOODTAUS","goodTaus"));
-  handler->addEventVariable("QGOODTAUS", new EventVariableObjectVariableVector<double>("CHARGE","goodTaus"));
+  handler->addEventVariable("QGOODTAUS", new EventVariableObjectVariableVector<int>("CHARGE","goodTaus"));
   handler->addEventVariable("PTGOODTAUS", new EventVariableObjectVariableVector<double>("PT","goodTaus"));
   handler->addEventVariable("ETAGOODTAUS", new EventVariableObjectVariableVector<double>("ETA","goodTaus"));
   handler->addEventVariable("fakeRoleGOODTAUS", new EventVariableObjectVariableVector<int>("fakeRole","goodTaus"));
@@ -956,32 +956,32 @@ void setupVariables2(BaseHandler* handler,bool isMC = false)
   handler->addEventVariable("ETAGOODPHOTONS", new EventVariableObjectVariableVector<double>("ETA","goodPhotons"));
 
   handler->addEventVariable("nINPELECTRONS", new EventVariableN("nINPELECTRONS","isoNonPromptElectrons"));
-  handler->addEventVariable("QINPELECTRONS", new EventVariableObjectVariableVector<double>("CHARGE","isoNonPromptElectrons"));
+  handler->addEventVariable("QINPELECTRONS", new EventVariableObjectVariableVector<int>("CHARGE","isoNonPromptElectrons"));
   handler->addEventVariable("PTINPELECTRONS", new EventVariableObjectVariableVector<double>("PT","isoNonPromptElectrons"));
   handler->addEventVariable("ETAINPELECTRONS", new EventVariableObjectVariableVector<double>("ETA","isoNonPromptElectrons"));
 
   handler->addEventVariable("nNIPELECTRONS", new EventVariableN("nNIPELECTRONS","promptNonIsoElectrons"));
-  handler->addEventVariable("QNIPELECTRONS", new EventVariableObjectVariableVector<double>("CHARGE","promptNonIsoElectrons"));
+  handler->addEventVariable("QNIPELECTRONS", new EventVariableObjectVariableVector<int>("CHARGE","promptNonIsoElectrons"));
   handler->addEventVariable("PTNIPELECTRONS", new EventVariableObjectVariableVector<double>("PT","promptNonIsoElectrons"));
   handler->addEventVariable("ETANIPELECTRONS", new EventVariableObjectVariableVector<double>("ETA","promptNonIsoElectrons"));
 
   handler->addEventVariable("nNINPELECTRONS", new EventVariableN("nNIPELECTRONS","nonPromptNonIsoElectrons"));
-  handler->addEventVariable("QNINPELECTRONS", new EventVariableObjectVariableVector<double>("CHARGE","nonPromptNonIsoElectrons"));
+  handler->addEventVariable("QNINPELECTRONS", new EventVariableObjectVariableVector<int>("CHARGE","nonPromptNonIsoElectrons"));
   handler->addEventVariable("PTNINPELECTRONS", new EventVariableObjectVariableVector<double>("PT","nonPromptNonIsoElectrons"));
   handler->addEventVariable("ETANINPELECTRONS", new EventVariableObjectVariableVector<double>("ETA","nonPromptNonIsoElectrons"));
   
   handler->addEventVariable("nINPMUONS", new EventVariableN("nINPMUONS","isoNonPromptMuons"));
-  handler->addEventVariable("QINPMUONS", new EventVariableObjectVariableVector<double>("CHARGE","isoNonPromptMuons"));
+  handler->addEventVariable("QINPMUONS", new EventVariableObjectVariableVector<int>("CHARGE","isoNonPromptMuons"));
   handler->addEventVariable("PTINPMUONS", new EventVariableObjectVariableVector<double>("PT","isoNonPromptMuons"));
   handler->addEventVariable("ETAINPMUONS", new EventVariableObjectVariableVector<double>("ETA","isoNonPromptMuons"));
   
   handler->addEventVariable("nNIPMUONS", new EventVariableN("nNIPMUONS","promptNonIsoMuons"));
-  handler->addEventVariable("QNIPMUONS", new EventVariableObjectVariableVector<double>("CHARGE","promptNonIsoMuons"));
+  handler->addEventVariable("QNIPMUONS", new EventVariableObjectVariableVector<int>("CHARGE","promptNonIsoMuons"));
   handler->addEventVariable("PTNIPMUONS", new EventVariableObjectVariableVector<double>("PT","promptNonIsoMuons"));
   handler->addEventVariable("ETANIPMUONS", new EventVariableObjectVariableVector<double>("ETA","promptNonIsoMuons"));
   
   handler->addEventVariable("nNINPMUONS", new EventVariableN("nNIPMUONS","nonPromptNonIsoMuons"));
-  handler->addEventVariable("QNINPMUONS", new EventVariableObjectVariableVector<double>("CHARGE","nonPromptNonIsoMuons"));
+  handler->addEventVariable("QNINPMUONS", new EventVariableObjectVariableVector<int>("CHARGE","nonPromptNonIsoMuons"));
   handler->addEventVariable("PTNINPMUONS", new EventVariableObjectVariableVector<double>("PT","nonPromptNonIsoMuons"));
   handler->addEventVariable("ETANINPMUONS", new EventVariableObjectVariableVector<double>("ETA","nonPromptNonIsoMuons"));
 
@@ -1003,7 +1003,7 @@ void setupVariables2(BaseHandler* handler,bool isMC = false)
   
   // Regular tracks
   handler->addEventVariable("NGOODTRACKS", new EventVariableN("NGOODTRACKS","goodTracks"));
-  handler->addEventVariable("QGOODTRACKS", new EventVariableObjectVariableVector<double>("CHARGE","goodTracks"));
+  handler->addEventVariable("QGOODTRACKS", new EventVariableObjectVariableVector<int>("CHARGE","goodTracks"));
   handler->addEventVariable("PTGOODTRACKS", new EventVariableObjectVariableVector<double>("PT","goodTracks"));
   handler->addEventVariable("ETAGOODTRACKS", new EventVariableObjectVariableVector<double>("ETA","goodTracks"));
